+++ src/main/kotlin/io/openfuture/chain/smartcontract/core/model/SmartContract.kt
@@ -1,13 +1,11 @@
 package io.openfuture.chain.smartcontract.core.model
 
 import io.openfuture.chain.smartcontract.core.exception.RequiredException
-import io.openfuture.chain.smartcontract.core.utils.AddressUtils
 
-abstract class SmartContract(ownerAddress: String) {
+abstract class SmartContract {
 
-    protected val owner: Address = Address(ownerAddress)
-    //TODO calculate from owner address (Keccak256 hash of sender address and nonce)
-    protected var address: Address = Address(AddressUtils.generateContractAddress(ownerAddress, "0"))
+    protected lateinit var owner: String
+    protected lateinit var address: String
 
 
     protected fun required(value: Boolean, message: String? = null) {