+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/exception/ContractExecutionException.kt
@@ -1,6 +1,6 @@
 package io.openfuture.chain.smartcontract.deploy.exception
 
-class ClassLoadingException(
-    message: String? = "Class loading failed",
+class ContractExecutionException(
+    message: String? = "Contract execution failed",
     cause: Throwable? = null
 ) : RuntimeException(message, cause)