+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/exception/ContractLoadingException.kt
@@ -1,0 +1,6 @@
+package io.openfuture.chain.smartcontract.deploy.exception
+
+class ContractLoadingException(
+    message: String? = "Class loading failed",
+    cause: Throwable? = null
+) : RuntimeException(message, cause)