+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/repository/ContractRepository.kt
@@ -1,0 +1,8 @@
+package io.openfuture.chain.smartcontract.deploy.repository
+
+import io.openfuture.chain.smartcontract.deploy.domain.ContractDto
+
+interface ContractRepository {
+
+    fun get(address: String): ContractDto
+}