+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/service/DefaultContractService.kt
@@ -1,24 +1,34 @@
 package io.openfuture.chain.smartcontract.deploy.service
 
 import io.openfuture.chain.smartcontract.deploy.domain.ClassSource
+import io.openfuture.chain.smartcontract.deploy.domain.ContractMethod
+import io.openfuture.chain.smartcontract.deploy.execution.ContractExecutor
 import io.openfuture.chain.smartcontract.deploy.load.SourceClassLoader
+import io.openfuture.chain.smartcontract.deploy.repository.ContractRepository
 import org.springframework.stereotype.Service
 
 @Service
-class DefaultContractService : ContractService {
+class DefaultContractService(
+    private val repository: ContractRepository,
+    private val executor: ContractExecutor
+) : ContractService {
 
     private val classLoader = SourceClassLoader()
 
-    
+
     override fun deploy(bytes: ByteArray) {
-        // simple deploy method
-        val source = ClassSource(bytes)
-        classLoader.loadBytes(source.qualifiedName, bytes)
+        classLoader.loadBytes(ClassSource(bytes))
     }
 
-    override fun run(className: String, method: String, vararg params: Any) {
-        //run a method of a contract in separate thread
-        TODO("not implemented")
+    //todo replace with transaction
+    override fun callMethod(contractAddress: String, methodName: String, vararg params: Any) {
+        val contact = repository.get(contractAddress)
+
+        //todo validateIsMethodExists()
+        val method = ContractMethod(methodName, params)
+
+        val result = executor.run(contact, method)
+        //todo update state
     }
 
 }