+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/service/Services.kt
@@ -4,6 +4,6 @@ interface ContractService {
 
     fun deploy(bytes: ByteArray)
 
-    fun run(className: String, method: String, vararg params: Any)
+    fun callMethod(contractAddress: String, methodName: String, vararg params: Any)
 
 }