+++ src/main/kotlin/io/openfuture/chain/smartcontract/deploy/validation/ValidationResult.kt
@@ -1,4 +1,4 @@
-package io.test.io.openfuture.chain.smartcontract.deploy.validation
+package io.openfuture.chain.smartcontract.deploy.validation
 
 class ValidationResult(
     val errors: MutableSet<String> = mutableSetOf()
@@ -10,6 +10,6 @@ class ValidationResult(
 
     fun hasErrors(): Boolean = errors.isNotEmpty()
 
-    override fun toString(): String = errors.joinToString(separator = "\n", prefix = "[", postfix = "]")
+    override fun toString(): String = errors.joinToString(separator = "\n", prefix = "\n", postfix = "\n")
 
 }